backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-openai
langchain-core
python-dotenv
httpx
sse-starlette
sqlalchemy[asyncio]>=2.0
asyncpg
pydantic-settings
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
